tweak the views to allow a home button
