+++ meal-ui/src/pages/ViewRecipe/View.js
@@ -1,12 +1,16 @@
 import RecipeCard from 'pages/ViewRecipe/RecipeCard';
 import { FormStatuses } from 'forms';
 import { demoImage } from 'DemoImage';
+import { HomeButton } from 'components/Button';
 
 export default function View(props) {
 	const { recipe, isLoading } = props;
 
 	return (
 		<div className='p-recipe-view'>
+			<div className='left-panel'>
+				<HomeButton />
+			</div>
 			<RecipeCard
 				recipe={recipe}
 				status={FormStatuses.Unknown}
